dbutils: warn on empty last_xact_replay_timestamp and dont try to log it
Under some (yet unclear) circumstances, last_xact_replay_timestamp can
be NULL, so repmgrd tries to insert NULL values into NON-NULL field in
its logging table, which causes it to fail and not going into operation.

This situation seems to be harmless, so just skip the failing logging
attempts in that case.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>